Remove `jsdom`, turns out we don't need it
The only reason it was "needed" before was that there was an unchecked access of
`window.location.search`, but `window` is undefined in the Node context, i.e.
when running tests.

But we actually don't need to access `window` when running tests, since the
tests are just testing pure logic. We can fix this by just guarding against
`window` being possibly undefined.